legacy/inference/requirements-asr.txt
accelerate==1.12.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
audioread==3.1.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
decorator==5.2.1
einops==0.8.2
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.1.0
gradio==6.5.1
gradio_client==2.0.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numba==0.63.1
onnxruntime-gpu==1.19.2
orjson==3.11.6
packaging==26.0
pandas==3.0.0
pillow==12.0.0
pip==25.3
platformdirs==4.5.1
pooch==1.8.2
protobuf==6.33.5
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.1
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
semantic-version==2.10.0
setuptools==65.5.0
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
sox==1.5.0
soxr==1.0.0
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
torchvision==0.19.0+cu124
tqdm==4.67.1
# transformers @ git+https://github.com/huggingface/transformers
typer==0.21.1
typer-slim==0.21.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
websockets==16.0
numpy==1.26.4